+++ src/common/consensus-provider/block-cache/block-cache.service.ts
@@ -2,9 +2,9 @@ import { LOGGER_PROVIDER } from '@lido-nestjs/logger';
 import { Inject, Injectable, LoggerService } from '@nestjs/common';
 
 import { ConfigService } from 'common/config';
+import { Epoch, RootHex, Slot } from 'common/types/types';
 
 import { BlockHeaderResponse, BlockInfoResponse } from '../intefaces';
-import { Epoch, RootHex, Slot } from '../types';
 
 export interface BlockCache {
   missed: boolean;
