--- src/common/consensus-provider/types/index.ts
@@ -1,1 +1,0 @@
-export * from './primitive.type';
