--- src/common/consensus-provider/types/primitive.type.ts
@@ -1,7 +1,0 @@
-export type BLSSignature = string;
-export type ValidatorIndex = string;
-export type RootHex = string;
-export type Slot = number;
-export type Epoch = number;
-export type BlockId = RootHex | Slot | 'head' | 'genesis' | 'finalized';
-export type StateId = RootHex | Slot | 'head' | 'genesis' | 'finalized' | 'justified';
