+++ src/duty/sync/sync.metrics.ts
@@ -4,10 +4,9 @@ import { Inject, Injectable, LoggerService } from '@nestjs/common';
 import { ConfigService } from 'common/config';
 import { allSettled } from 'common/functions/allSettled';
 import { PrometheusService, TrackTask, setOtherOperatorsMetric, setUserOperatorsMetric } from 'common/prometheus';
+import { Epoch } from 'common/types/types';
 import { ClickhouseService } from 'storage';
 import { RegistryService, RegistrySourceOperator } from 'validators-registry';
-
-import { Epoch } from '../../common/consensus-provider/types';
 
 @Injectable()
 export class SyncMetrics {
