+++ src/duty/sync/sync.service.ts
@@ -4,8 +4,8 @@ import { Inject, Injectable, LoggerService } from '@nestjs/common';
 
 import { ConfigService } from 'common/config';
 import { BlockInfoResponse, ConsensusProviderService, SyncCommitteeValidator } from 'common/consensus-provider';
-import { Epoch, Slot, StateId } from 'common/consensus-provider/types';
 import { PrometheusService, TrackTask } from 'common/prometheus';
+import { Epoch, Slot, StateId } from 'common/types/types';
 import { SummaryService } from 'duty/summary';
 
 import { SYNC_COMMITTEE_SIZE } from './sync.constants';
