+++ src/inspector/inspector.service.ts
@@ -5,9 +5,9 @@ import { CriticalAlertsService } from 'common/alertmanager';
 import { ConfigService, WorkingMode } from 'common/config';
 import { BlockHeaderResponse, ConsensusProviderService } from 'common/consensus-provider';
 import { BlockCacheService } from 'common/consensus-provider/block-cache';
-import { Slot } from 'common/consensus-provider/types';
 import { sleep } from 'common/functions/sleep';
 import { PrometheusService, TrackTask } from 'common/prometheus';
+import { Slot } from 'common/types/types';
 import { DutyMetrics, DutyService } from 'duty';
 import { ClickhouseService } from 'storage';
 import { EpochProcessingState } from 'storage/clickhouse';
